+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Respect client capabilities for snippet support in completion items
 - Respect `luau-lsp.completion.addParentheses` when `fillCallArguments` is enabled
 - Fixed Inlay Hints crash when calling a function which only takes varargs
+- Fixed Request Failed due to "No managed text document" as URLs were not being updated correctly
 
 ## [1.13.0] - 2022-10-28